+++ src/Mailman.Server/ClientApp/src/App.js
@@ -3,8 +3,8 @@ import { Route } from 'react-router';
 import Layout from './components/Layout';
 import Home from './components/Home';
 import Settings from './components/Settings';
-import { TitlePageContainer } from './components/MergeTemplate/TitlePageContainer';
-import { TabSelectionContainer } from './components/MergeTemplate/TabSelectionContainer';
+import { TitlePageContainer } from './components/TitlePageContainer';
+import { TabSelectionContainer } from './components/TabSelectionContainer';
 import Counter from './components/Counter';
 import FetchData from './components/FetchData';
