+++ src/Mailman.Server/ClientApp/src/components/TitlePageContainer.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 
 import TitlePage from './TitlePage';
-import { updateTitlePage } from '../../actions/createMergeTemplate'
+import { updateTitlePage } from '../actions/createMergeTemplate'
 
 function mapStateToProps(state) {
     const { currentMergeTemplate } = state;
